Fix: separate sanitizeGpxFileNames from processGpxFiles in different js files and in actions
Separate `sanitizeGpxFileNames` from `processGpxFiles` into different JavaScript files and update GitHub Actions workflows.

* **New File**: Add `scripts/sanitize-gpx.js` to contain the `sanitizeGpxFileNames` function.
* **Process GPX Script**: Modify `scripts/process-gpx.js` to remove the `sanitizeGpxFileNames` function and import it from `scripts/sanitize-gpx.js`.
* **GitHub Actions**: Update `.github/workflows/deploy.yml` and `.github/workflows/verify-pr.yml` to run `scripts/sanitize-gpx.js` before `scripts/process-gpx.js`.
* **Documentation**: Update `README.md` to reflect the separation of `sanitizeGpxFileNames` and `processGpxFiles` into separate scripts.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/statnmap/gpx-traces-website?shareId=XXXX-XXXX-XXXX-XXXX).
