+++ .github/workflows/deploy.yml
@@ -21,6 +21,9 @@ jobs:
       - name: Install dependencies
         run: npm install
 
+      - name: Run sanitize script
+        run: node scripts/sanitize-gpx.js
+
       - name: Run pre-build script
         run: node scripts/process-gpx.js
