+++ .github/workflows/verify-pr.yml
@@ -24,6 +24,9 @@ jobs:
       - name: Check Dependabot alerts
         run: npm audit
 
+      - name: Run sanitize script
+        run: node scripts/sanitize-gpx.js
+
       - name: Run pre-build script
         run: node scripts/process-gpx.js
